Revert "Misc"
This reverts commit e9b81b66fa755949fb894f7c7b9afc992ba96df3.

+++ src/internal.h
@@ -259,7 +259,7 @@ typedef struct {
 	module *curr_db;
 	node *curr_term, *curr_match;
 	uint32_t env_point, curr_trail, prev_choice, curr_frame;
-	uint16_t frame_size, trail_size;
+	uint8_t frame_size, trail_size;
 
 	struct {
 		unsigned cut : 1;
